Bug Fixes and Functionality added
Now breaks will overlap in a better way, before due to this the schedule would freeze if more than 12 people were wokring while having break

Added the functionality that allows the user to generate a schedule with or without runner. Useful if one will have a runner that will not be zoned anywhere else.